Remove unreachable code path
In `small_approximation.js`, the check for `x === 0` was removed as
as this path should never be reached. In the original Cephes
implementation, the check was needed as `x == 0` was possible. In
this implementation, however, we check for `x == 0` as a special
condition, preventing it from triggering later code paths. Note that
the original also does not take into account negative zero, which
would be a bug in JavaScript.