Remove unreachable code path
In `index.js`, a code path could only be reached when `sin(v)`
evaluated to zero. This occurrence is not possible. For positive
integers > 33, the implementation uses Stirling's approximation.
For negative integers < -33, the implementation includes a check
which returns `NaN`. Accordingly, the path in question could only be
feasibly triggered when abs(x)-floor(abs(x)) is approx 0 or approx
1. When approx 1, the difference (delta) is reflected so to be
approx equal to 0. Accordingly, `sin(pi*delta)` can only be zero
when `pi*delta` evaluates to `0`, which should never happen, as
delta can never be less than machine epsilon, without `x` being
considered a negative integer and triggering an alternative code
pathway. The only other way for the pathway to be triggered is due
to a poor sine implementation which returns `0` when its argument
is `1e-15`.